Avoid importing jmp cli when running j command
