+++ packages/jumpstarter-cli/jumpstarter_cli/__main__.py
@@ -1,6 +1,6 @@
 """Allow running Jumpstarter through `python -m jumpstarter_cli`."""
 
-from . import jmp
+from .jmp import jmp
 
 if __name__ == "__main__":
     jmp(prog_name="jmp")
